codes/requirements.txt
django
django-phonenumber-field
phonenumbers
django-ses
djangorestframework
boto3
requests
django-crispy-forms
Pillow
psycopg2-binary
gunicorn
newrelic
django-allauth
#django-rest-auth
django-cors-headers
stripe
django-magic-link
drf_yasg
django-dbbackup
slack
pandas
djangorestframework-simplejwt
django-allauth
pytube
django-extensions
Werkzeug
